wifi: Always track interference when not locking on PPDU
This commit ensures the PPDU that is being received and we cannot lock
on is added as interference. Previously, this did not happen when the
PPDU was too weak to process, or when it was using a modulation that was
not supported.

This commit also de-comments TestUnsupportedModulationReception in
the wifi-phy-reception test suite, which failed before this commit and
does not fail on this commit.