Internet: Functionality to get Node from IP Address
This patch introduces new public APIs called `Ipv4AddressHelper::GetNode ()`
and `Ipv6AddressHelper::GetNode ()` for finding the `Ptr<Node>` for a given
`Ipv4Address` and `Ipv6Address` respectively. The inputs to this API
are `Ipv4Address`/`Ipv6Address` and `NodeContainer` to find the `Ptr<Node>`.

We introduce a function called `IsGlobalUnicast ()` in `Ipv4Address` and
`Ipv6Address`.

In `Ipv4Address`, it checks if the address is not a Broadcast, Localhost,
Multicast or Any address.

In `Ipv6Address`, it checks if the address starts with a 2 or 3, and is not
a documentation address.

We use these functions in the respective `GetNode ()` function and return
a nullptr if `IsGlobalUnicast ()` returns false.

For `Ipv4Address`, `GetNode ()` function will return a node pointer for valid
unicast IPv4 address on a non-loopback interface only.

For `Ipv6Address`, `GetNode ()` function will return a node pointer for valid
global unicast IPv6 address.

If `NodeContainer` is not provided, it finds the `Ptr<Node>` from all the
nodes present in the simulation.

Signed-off-by: Ameya Deshpande <ameyanrd@outlook.com>